Struct edcert::certificate_validator::CertificateValidator

pub struct CertificateValidator<R: Revoker> {
    // some fields omitted
}

This struct can be used to validate Certificates.

Methods

impl<R: Revoker> CertificateValidator<R>

fn new(master_public_key: &[u8; 32], revoker: R) -> CertificateValidator<R>

Call this to create a CV with a revoke server. For every certificate the revoke server is asked if it is known.

fn is_valid<V: Validatable>(&self, cert: &V) -> Result<(), &'static str>

Checks the certificate if it is valid. If the CV knows a revoke server, that is queried as well.

fn is_revoked<V: Validatable>(&self, cert: &V) -> Result<(), &'static str>