Trait edcert::certificate_validator::Validatable

pub trait Validatable {
    fn is_valid(&self, &[u8]) -> Result<(), &'static str>;
    fn is_revokable(&self) -> bool;
    fn get_id(&self) -> String;
}

This trait is implemented for types, which must be validated.

Required Methods

fn is_valid(&self, &[u8]) -> Result<(), &'static str>

This method is given the master public key and should check, if it is valid given that key.

fn is_revokable(&self) -> bool

This method should return true iff the object can be revoked. That is true for a public key, but not for a signature.

fn get_id(&self) -> String

If is_revokable() returns true, you must implement this function. It should return a representation by which the object can be identified by the revoke-list or revoke server.