Struct edcert::certificate_verificator::CertificateVerificator

pub struct CertificateVerificator {
    // some fields omitted
}

Methods

impl CertificateVerificator

fn new(master_public_key: &[u8; 32]) -> CertificateVerificator

Call this to create a CV without a revoke server.

fn with_revokeserver(revokeserver: &str, master_public_key: &[u8; 32]) -> CertificateVerificator

Call this to create a CV with a revoke server. For every certificate the revoke server is asked if it is known.

fn is_valid(&self, cert: &Certificate) -> Result<(), &'static str>

Checks the certificate if it is valid. If the CV knows a revoke server, that is queried as well.