Struct edcert::meta::Meta

pub struct Meta {
    // some fields omitted
}

Methods

impl Meta

fn new_empty() -> Meta

Creates a new Meta object, which can be used to store metadata for a certificate.

fn new(values: BTreeMap<String, String>) -> Meta

Creates a new Meta object using the given BTreeMap.

fn key_exists(&self, key: &str) -> bool

This method returns true iff the key exists.

fn get(&self, key: &str) -> Option<&String>

This method returns the value of key, if it exists. Otherwise it returns None.

fn set(&mut self, key: &str, value: &str)

This method assigns a value to a given key.

fn get_values_mut(&mut self) -> &mut BTreeMap<String, String>

This method returns a mutable reference to the tree object.

fn get_values(&self) -> &BTreeMap<String, String>

This method returns a reference to the tree object.

fn fill_bytes(&self, bytes: &mut [u8])

This method fills the given byte vector with a "hash" which is created from all keys and values.