Struct edcert::signature::Signature

pub struct Signature {
    // some fields omitted
}

Methods

impl Signature

fn new(signature: Vec<u8>) -> Signature

Creates a new Signature with the given signature. It is assumed that the signature is computed usign the master key.

fn with_parent(parent: Box<Certificate>, signature: Vec<u8>) -> Signature

Creates a new Signature with the given parent and given signature.

fn is_signed_by_master(&self) -> bool

This method will return true iff the certificate has no parent certificate. It is then signed with the master key.

fn get_parent(&self) -> Option<&Certificate>

This method will return the parent Certificate, or None, if it is signed with the master key.

fn get_hash(&self) -> &Vec<u8>

This method will return the signature given by the parent.